[DEV-8580] minor refactors
